relocate ens into services folder
