+++ packages/sdk/src/actions/swap/namespaces/BaseSwapNamespace.ts
@@ -1,7 +1,11 @@
 import { BaseNamespace } from '@/actions/shared/BaseNamespace.js'
 import type { SwapProvider } from '@/actions/swap/core/SwapProvider.js'
 import type { SupportedChainId } from '@/constants/supportedChains.js'
-import type { SwapQuoteParamsResolved } from '@/ens/types.js'
+import type { SwapQuoteParamsResolved } from '@/services/ens/types.js'
+import {
+  passthroughResolver,
+  type RecipientResolver,
+} from '@/services/ens/utils.js'
 import type { SwapProviderName, SwapSettings } from '@/types/actions.js'
 import type { Asset } from '@/types/asset.js'
 import type {
@@ -13,7 +17,6 @@ import type {
   SwapQuote,
   SwapQuoteParams,
 } from '@/types/swap/index.js'
-import { passthroughResolver, type RecipientResolver } from '@/utils/ens.js'
 
 type ConfiguredSwapProvider = SwapProvider<SwapProviderConfig>
