+++ packages/sdk/src/services/ens/utils.ts
@@ -6,10 +6,10 @@ import {
   EnsNotConfiguredError,
   EnsResolutionError,
   EnsRpcError,
-} from '@/ens/errors.js'
-import type { EnsName } from '@/ens/types.js'
+} from '@/services/ens/errors.js'
+import type { EnsName } from '@/services/ens/types.js'
 
-export { type EnsName, isEnsName } from '@/ens/types.js'
+export { type EnsName, isEnsName } from '@/services/ens/types.js'
 
 /** Resolves an ENS name or address to a checksummed hex Address, or returns undefined. */
 export type RecipientResolver = (
