+++ packages/sdk/src/wallet/core/wallets/abstract/Wallet.ts
@@ -3,8 +3,8 @@ import type { Address, LocalAccount } from 'viem'
 import { WalletLendNamespace } from '@/actions/lend/namespaces/WalletLendNamespace.js'
 import { WalletSwapNamespace } from '@/actions/swap/namespaces/WalletSwapNamespace.js'
 import type { SupportedChainId } from '@/constants/supportedChains.js'
-import { EnsNamespace } from '@/ens/index.js'
 import type { ChainManager } from '@/services/ChainManager.js'
+import { EnsNamespace } from '@/services/ens/index.js'
 import { fetchERC20Balance, fetchETHBalance } from '@/services/tokenBalance.js'
 import type { SwapSettings } from '@/types/actions.js'
 import type { Asset, TokenBalance } from '@/types/asset.js'
